testnet setup

+++ lib/network.py
@@ -950,15 +950,18 @@ class Network(util.DaemonThread):
     def init_headers_file(self):
         b = self.blockchains[0]
         filename = b.path()
-        open(filename, 'wb+').close()
-        self.downloading_headers = False
-        return
-        
-        print(b.get_hash(0), NetworkConstants.GENESIS)
-        if b.get_hash(0) == NetworkConstants.GENESIS:
+
+        if NetworkConstants.TESTNET:
+            if not os.path.exists(filename) or b.get_hash(0) != NetworkConstants.GENESIS:
+                open(filename, 'wb+').close()
+            with b.lock: b.update_size()
             self.downloading_headers = False
             return
         
+        if b.get_hash(0) == NetworkConstants.GENESIS:
+            self.downloading_headers = False
+            return
+
         def download_thread():
             try:
                 import urllib, socket
@@ -981,7 +984,6 @@ class Network(util.DaemonThread):
         t.start()
 
     def run(self):
-        print('zzzzzzzzzzzzzzzzzzzz')
         self.init_headers_file()
         while self.is_running() and self.downloading_headers:
             time.sleep(1)
